moved methods into their respective classes (refactoring)
