+++ src/main/java/services/FaceitMatchService.java
@@ -1,6 +1,7 @@
 package services;
 
 import com.google.gson.*;
+import http.ConnectionBuilder;
 import model.faceit.FaceitMatch;
 import net.dv8tion.jda.api.entities.Guild;
 import spark.Request;
@@ -18,10 +19,13 @@ public class FaceitMatchService {
     private Properties properties;
     ResourceBundle resourceBundle;
     DataService dataService;
+    ConnectionBuilder connectionBuilder;
+
 
     public FaceitMatchService(Properties properties, DataService dataService) {
         this.properties = properties;
         this.dataService = dataService;
+        connectionBuilder = new ConnectionBuilder(properties);
     }
 
     public void handleFaceitMatchStartEvent(Request request, List<Guild> allGuilds) {
@@ -35,10 +39,12 @@ public class FaceitMatchService {
         System.out.println("-------");
         String userId = request.params("user");
         try {
-            FaceitMatch match = fetchMatchDetails(userId);
+            FaceitMatch match = connectionBuilder.fetchFaceitMatchDetails(userId);
             System.out.println(match.getMatch_id()); // placeholder
-        } catch (IOException | InterruptedException e) {
-            System.out.println("Error while fetching Match Info...");
+        } catch (IOException ex) {
+            System.out.println("IOException thrown: " + ex.getMessage());
+        } catch (InterruptedException ex) {
+            System.out.println("InterruptedException thrown: " + ex.getMessage());
         }
     }
 
@@ -53,49 +59,5 @@ public class FaceitMatchService {
         System.out.println("-------");
 
         //TODO use the data service
-    }
-
-    private FaceitMatch fetchMatchDetails(String userId) throws IOException, InterruptedException {
-        HttpClient client = HttpClient.newHttpClient();
-
-        String matchId = fetchMatchId(client, userId);
-
-        HttpRequest request;
-        FaceitMatch responseData;
-
-        request = HttpRequest.newBuilder()
-                .uri(URI.create("https://open.faceit.com/data/v4/matches/" + matchId))
-                .header("Authorization", "Bearer " + properties.getProperty("faceit.api"))
-                .build();
-        HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
-
-        responseData = new Gson().fromJson(response.body(), FaceitMatch.class);
-
-        return responseData;
-    }
-
-    private String fetchMatchId(HttpClient client, String userId) throws IOException, InterruptedException {
-        HttpRequest request;
-        String id = null;
-
-        request = HttpRequest.newBuilder()
-                .uri(URI.create("https://www.faceit.com/api/match/v1/matches/groupByState?userId=" + userId))
-                .build();
-        HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
-
-        JsonElement jsonElement = JsonParser.parseString(response.body());
-        JsonObject jsonObject = jsonElement.getAsJsonObject();
-
-        JsonObject payload = jsonObject.getAsJsonObject("payload");
-        JsonArray ongoingArray = payload.getAsJsonArray("ONGOING");
-        JsonArray readyArray = payload.getAsJsonArray("READY");
-        if (ongoingArray != null && !ongoingArray.isEmpty()) {
-            JsonObject firstOngoingItem = ongoingArray.get(0).getAsJsonObject();
-            id = firstOngoingItem.get("id").getAsString();
-        } else if (readyArray != null  && !readyArray.isEmpty()) {
-            JsonObject firstReadyItem = readyArray.get(0).getAsJsonObject();
-            id = firstReadyItem.get("id").getAsString();
-        }
-        return id;
     }
 }
